+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed 
 To create a better separation between the generic 4-Corner datamodel classes and the SMP client related code the project was split into two projects.
-This projects now only contains generic code not related to specific functionality. The SMP client code has been moved to a new project:
+This projects now only contains generic code not related to specific functionality. The SMP client code has been moved to a new project: 
+[BDXR SMP Client](https://github.com/holodeck-b2b/bdxr-smp-client)
 
 ### Removed
 * The possibility to include multiple process identifiers in `org.holodeckb2b.bdxr.smp.datamodel.ProcessInfo` as this is not